refactor: move ta assessment task route  to assessment routes
